Add ydt-parametrized KDA; verified activation parity with YDT
